Introduce a domain layer for the score
I went for a revers approach, in which I implemented first the infrastructure, then the application layer and then the domain layer. This was due to heavy problems in finding a owned domain which doesn't require replicating the game state. Still doesn't carry strong references to the players which is a wip